+++ tests/contrib/flask/flask_tests.py
@@ -258,6 +258,34 @@ def test_options_request(flask_apm_client):
     assert transactions[0]["context"]["request"]["method"] == "OPTIONS"
 
 
+@pytest.mark.parametrize(
+    "elasticapm_client", [{"capture_headers": "true"}, {"capture_headers": "false"}], indirect=True
+)
+def test_capture_headers_errors(flask_apm_client):
+    resp = flask_apm_client.app.test_client().post("/an-error/", headers={"some-header": "foo"})
+    resp.close()
+    error = flask_apm_client.client.events[ERROR][0]
+    if flask_apm_client.client.config.capture_headers:
+        assert error["context"]["request"]["headers"]["some-header"] == "foo"
+    else:
+        assert "headers" not in error["context"]["request"]
+
+
+@pytest.mark.parametrize(
+    "elasticapm_client", [{"capture_headers": "true"}, {"capture_headers": "false"}], indirect=True
+)
+def test_capture_headers_transactions(flask_apm_client):
+    resp = flask_apm_client.app.test_client().post("/users/", headers={"some-header": "foo"})
+    resp.close()
+    transaction = flask_apm_client.client.events[TRANSACTION][0]
+    if flask_apm_client.client.config.capture_headers:
+        assert transaction["context"]["request"]["headers"]["some-header"] == "foo"
+        assert transaction["context"]["response"]["headers"]["foo"] == "bar;baz"
+    else:
+        assert "headers" not in transaction["context"]["request"]
+        assert "headers" not in transaction["context"]["response"]
+
+
 def test_streaming_response(flask_apm_client):
     resp = flask_apm_client.app.test_client().get("/streaming/")
     assert resp.data == b"01234"